+++ partial-recall.el
@@ -122,6 +122,11 @@ This is will implant buffers that have met
                  (const :tag "None" nil))
   :group 'partial-recall)
 
+(defcustom partial-recall-filter '("COMMIT_EDITMSG")
+  "Names of buffers that should be ignored."
+  :type '(repeat regexp)
+  :group 'partial-recall)
+
 ;;; -- Internal variables
 
 (defvar partial-recall--table (make-hash-table))
@@ -760,7 +765,9 @@ the max age."
 
 (defun partial-recall--meaningful-buffer-p (buffer)
   "Check if BUFFER should be remembered."
-  (not (null (buffer-file-name buffer))))
+  (and-let* ((name (buffer-file-name buffer))
+             (filter (mapconcat (lambda (it) (concat "\\(?:" it "\\)")) partial-recall-filter "\\|"))
+             ((not (string-match-p filter (buffer-name buffer)))))))
 
 ;;; -- Utility